add task descriptions and nicer logs for tasks (#491)
* task descriptions and nicer logs for tasks

* add more parametrisation

* field descriptions

* fix startup and add makefile

* remove runtask command

+++ app/tasks/layer2_import_designation.py
@@ -19,11 +19,15 @@ class Layer2ImportDesignationTask(interface.Task):
         batch_size: int = 100000,
         dry_run: bool = False,
         silent: bool = False,
+        since: datetime.datetime | str | None = None,
+        cleanup_orphans: bool = True,
     ) -> None:
         self.log = logger
         self.batch_size = batch_size
         self.dry_run = dry_run
         self.silent = silent
+        self.since = interface.parse_since(since)
+        self.cleanup_orphans = cleanup_orphans
 
     @classmethod
     def name(cls) -> str:
@@ -36,11 +40,15 @@ class Layer2ImportDesignationTask(interface.Task):
         self.layer2_repository = repositories.Layer2Repository(self.pg_storage, self.log)
 
     def run(self) -> None:
-        last_update_dt = self.layer2_repository.get_last_update_time(model.RawCatalog.DESIGNATION)
+        if self.since is not None:
+            last_update_dt = self.since
+        else:
+            last_update_dt = self.layer2_repository.get_last_update_time(model.RawCatalog.DESIGNATION)
         self.log.info(
             "Starting Layer 2 designation import",
             last_update=last_update_dt.ctime(),
             dry_run=self.dry_run,
+            cleanup_orphans=self.cleanup_orphans,
         )
 
         objects_to_save = 0
@@ -77,11 +85,13 @@ class Layer2ImportDesignationTask(interface.Task):
                 total_processed=objects_to_save,
             )
 
-        orphaned = self.layer2_repository.get_orphaned_pgcs([model.RawCatalog.DESIGNATION])
-        pgcs_to_remove = [pgc for pgcs in orphaned.values() for pgc in pgcs]
-        orphans_to_delete = len(pgcs_to_remove)
-        if pgcs_to_remove and not self.dry_run:
-            self.layer2_repository.remove_pgcs([model.RawCatalog.DESIGNATION], pgcs_to_remove)
+        orphans_to_delete = 0
+        if self.cleanup_orphans:
+            orphaned = self.layer2_repository.get_orphaned_pgcs([model.RawCatalog.DESIGNATION])
+            pgcs_to_remove = [pgc for pgcs in orphaned.values() for pgc in pgcs]
+            orphans_to_delete = len(pgcs_to_remove)
+            if pgcs_to_remove and not self.dry_run:
+                self.layer2_repository.remove_pgcs([model.RawCatalog.DESIGNATION], pgcs_to_remove)
 
         if not self.dry_run:
             self.layer2_repository.update_last_update_time(
